third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/form-controls/input-line-height-computed-expected.txt

This is a testharness.js-based test.
[FAIL] getComputedStyle(<input type=text>).lineHeight should return a used value that is no smaller than 'normal' (but should not literally be 'normal')
  assert_not_equals: usedLineHeight got disallowed value "1px"
[PASS] <input type=text>.computedStyleMap().get('line-height') should not be affected by the used value clamping
[FAIL] getComputedStyle(<input type=tel>).lineHeight should return a used value that is no smaller than 'normal' (but should not literally be 'normal')
  assert_not_equals: usedLineHeight got disallowed value "1px"
[PASS] <input type=tel>.computedStyleMap().get('line-height') should not be affected by the used value clamping
[FAIL] getComputedStyle(<input type=search>).lineHeight should return a used value that is no smaller than 'normal' (but should not literally be 'normal')
  assert_not_equals: usedLineHeight got disallowed value "1px"
[PASS] <input type=search>.computedStyleMap().get('line-height') should not be affected by the used value clamping
[FAIL] getComputedStyle(<input type=url>).lineHeight should return a used value that is no smaller than 'normal' (but should not literally be 'normal')
  assert_not_equals: usedLineHeight got disallowed value "1px"
[PASS] <input type=url>.computedStyleMap().get('line-height') should not be affected by the used value clamping
[FAIL] getComputedStyle(<input type=email>).lineHeight should return a used value that is no smaller than 'normal' (but should not literally be 'normal')
  assert_not_equals: usedLineHeight got disallowed value "1px"
[PASS] <input type=email>.computedStyleMap().get('line-height') should not be affected by the used value clamping
[FAIL] getComputedStyle(<input type=password>).lineHeight should return a used value that is no smaller than 'normal' (but should not literally be 'normal')
  assert_not_equals: usedLineHeight got disallowed value "1px"
[PASS] <input type=password>.computedStyleMap().get('line-height') should not be affected by the used value clamping
Harness: the test ran to completion.
